esversion: 2016 (#3060)
* [[CHORE]] Remove unreachable branch

The `inES5` method is never invoked with a parameter, so the branch in
question is never executed. Remove the parameter, the parameter's
description, and the associated branch.

* [[CHORE]] Refactor to support `esversion: 2016`

This change simply enables the configuration settings `esversion: 7` and
`esversion: 2016` without introducing any new parsing/linting
functionality.

* [[FEAT]] Introduce exponentiation operator

* [[FEAT]] Enforce ES2016 restriction on USD

ECMAScript 2016 introduced an early error for any function situated in
non-strict code whose parameter list is non-simple and whose body
contains a "use strict" directive. Enforce this restriction when
interpreting input according to that revision of the specification.